Translate contact form fields to Portuguese
The contact form labels, placeholders and success message were translated to Portuguese to improve user experience for Portuguese-speaking clients. This was achieved by changing the hard-coded English strings to their Portuguese equivalent on the contact view component. The contact view component was uncommented on the page component to be displayed to users.